Add the index homepage and the chili recipe page
The initial project only had a README file. Create an index.html homepage, a recipes directory to store all recipe pages, and a chili.html page within the recipes directory.
Provide boilerplates for both the index and chili HTML pages. Include links in both the index and chili pages that navigate the user from the homepage, to the chili recipe page, and back to the homepage.
The index homepage and the chili recipe page now have a basic HTML document structure with additional links to each other.