Add cereal recipe page with content and cereal image to the project
The previous project had 2 recipe pages. Create a cereal.html recipe page
within the recipes directory. Add a froot_loops_bowl.jpg image in the
images directory. Include a boilerplate in cereal page with a link to the
index.html homepage and in that homepage include a link to the cereal page.
Provide the main content in the cereal page including the froot loops image,
recipe description, ingredients, and steps. Remove relationship attributes
for all links in the project because they use self targets. The project
now has three functional recipe pages linked to the homepage.